Port connectivity from 10.54 kernel

+++ drivers/misc/mediatek/connectivity/combo/common/core/wmt_ic_6628.c
@@ -660,8 +660,8 @@ static INT32 mt6628_sw_init(P_WMT_HIF_CONF pWmtHifConf)
 	INT32 iRet = -1;
 	UINT32 u4Res = 0;
 	UINT8 evtBuf[256];
-	unsigned long ctrlPa1;
-	unsigned long ctrlPa2;
+	UINT32 ctrlPa1;
+	UINT32 ctrlPa2;
 	UINT32 hw_ver;
 #if CFG_WMT_MULTI_PATCH
 	UINT32 patch_num = 0;
@@ -1170,8 +1170,8 @@ static INT32 mt6628_ver_check(VOID)
 	UINT32 fw_ver;
 	INT32 iret;
 	const WMT_IC_INFO_S *p_info;
-	unsigned long ctrlPa1;
-	unsigned long ctrlPa2;
+	UINT32 ctrlPa1;
+	UINT32 ctrlPa2;
 
 	/* 1. identify chip versions: HVR(HW_VER) and FVR(FW_VER) */
 	WMT_LOUD_FUNC("MT6628: before read hw_ver (hw version)\n");
@@ -1266,7 +1266,7 @@ static const WMT_IC_INFO_S *mt6628_find_wmt_ic_info(const UINT32 hw_ver)
 static INT32 wmt_stp_init_coex(VOID)
 {
 	INT32 iRet;
-	unsigned long addr;
+	size_t addr;
 	WMT_GEN_CONF *pWmtGenConf;
 
 #define COEX_WMT  0
@@ -1383,7 +1383,7 @@ static INT32 mt6628_set_sdio_driving(VOID)
 {
 	INT32 ret = 0;
 
-	unsigned long addr;
+	size_t addr;
 	WMT_GEN_CONF *pWmtGenConf;
 	UINT32 drv_val = 0;
